requirements.txt
yt-dlp
demucs
pillow
shutup
tqdm
selenium
